api/app/requirements.txt
Flask
numpy
pandas
scikit-learn
mysql-connector-python